use search range directly
instead of searching all Chunk and filtering results afterwards
