+++ search/searchentityplugin.h
@@ -19,7 +19,7 @@ class SearchEntityPlugin : public QWidget, public SearchPluginI
 
   QWidget &getWidget() override;
 
-  SearchPluginI::ResultListT searchChunk(const Chunk &chunk) override;
+  SearchPluginI::ResultListT searchChunk(const Chunk &chunk, const Range<int> &range) override;
 
  private:
   QLayout* layout;
